fix(justfile, overlay): add `dir` type for local overrides

+++ Justfile
@@ -34,25 +34,39 @@ tree IMAGE="localhost/bluefin-common:latest":
     rm TreeContainerfile
     podman rmi tree-temp
 
-overlay $BLUEFIN_MERGE="1":
+overlay $BLUEFIN_MERGE="1" $SOURCE="dir":
     #!/usr/bin/env bash
     ROOTFS_DIR="$(mktemp -d --tmpdir="${ROOTFS_BASE:-/tmp}")"
     trap 'rm -rf "${ROOTFS_DIR}"' EXIT
     NAME_TRIMMED=bfincommon
-    crane export ghcr.io/projectbluefin/common:latest - | tar -xvf - -C "${ROOTFS_DIR}"
+
+    if [ "$SOURCE" == "dir" ] ; then
+        cp -a ./system_files/shared/. "${ROOTFS_DIR}"
+        if [ "${BLUEFIN_MERGE}" == "1" ] ; then
+            cp -a ./system_files/bluefin/. "${ROOTFS_DIR}"
+        fi
+    elif [ "$SOURCE" == "image" ] ; then
+        podman export "$(podman create ghcr.io/projectbluefin/common:latest)" -o - | tar -xvf - -C "${ROOTFS_DIR}"
+    fi
+
     install -d -m0755 "${ROOTFS_DIR}/usr/lib/extension-release.d"
     tee "${ROOTFS_DIR}/usr/lib/extension-release.d/extension-release.${NAME_TRIMMED}" <<EOF
     ID="_any"
     ARCHITECTURE="$(sed 's/_/-/g' <<< "$(arch)")"
     EOF
-    cp -a "${ROOTFS_DIR}/system_files/shared/." "${ROOTFS_DIR}"
-    if [ "${BLUEFIN_MERGE}" == "1" ] ; then
-        cp -a "${ROOTFS_DIR}/system_files/bluefin/." "${ROOTFS_DIR}"
+
+    if [ -e "${ROOTFS_DIR}/system_files" ] ; then 
+        cp -a "${ROOTFS_DIR}/system_files/shared/." "${ROOTFS_DIR}"
+        if [ "${BLUEFIN_MERGE}" == "1" ] ; then
+            cp -a "${ROOTFS_DIR}/system_files/bluefin/." "${ROOTFS_DIR}"
+        fi
+        rm -r "${ROOTFS_DIR}/system_files"
     fi
-    rm -r "${ROOTFS_DIR}/system_files"
+
     if [ -d "${ROOTFS_DIR}/etc" ] ; then
         mv --no-clobber "${ROOTFS_DIR}/etc" "${ROOTFS_DIR}/usr/etc"
     fi
+
     for dir in "var" "run"; do
         if [ -d "${ROOTFS_DIR}"/"${dir}" ] ; then
             rm -r "${ROOTFS_DIR:?}/${dir}"
